third_party/blink/web_tests/http/tests/mime/javascript-mimetype-usecounters-expected.txt

ALERT: Correct
ALERT: Correct
ALERT: Correct
ALERT: Correct
CONSOLE ERROR: Refused to execute script from 'http://127.0.0.1:8000/mime/resources/javascript-mimetype.php?mimetype=image/foo' because its MIME type ('image/foo') is not executable.
CONSOLE ERROR: Refused to execute script from 'http://localhost:8000/mime/resources/javascript-mimetype.php?mimetype=image/bar' because its MIME type ('image/bar') is not executable.
ALERT: Correct
ALERT: Correct
ALERT: Correct
ALERT: Correct
ALERT: Correct
ALERT: Correct
ALERT: Correct
ALERT: Correct
ALERT: Correct
ALERT: Correct
